fix(recurrence): fix Numba overflow, edge cases, cache keys; strengthen tests
- Fix int32 overflow in Numba diagonal scanning for n > 46,340 (use int64)
- Guard estimate_tau against empty TDMI arrays (max_shift=0)
- Include **kwargs in cache keys for estimate_tau and estimate_embedding_dim
- Validate theiler_window is non-negative
- Rewrite ever-passing cache tests to verify actual cache behavior
- Tighten weak DET assertions and call get_spectrum instead of hasattr
- Document from_adjacency None attributes

+++ src/driada/recurrence/_numba_rqa.py
@@ -23,8 +23,9 @@ try:
         diag_idx = cols - rows
         positions = rows
 
-        # Sort by (diag_idx, position)
-        order = np.argsort(diag_idx * n + positions)
+        # Sort by (diag_idx, position) — use int64 to avoid overflow
+        # when n > 46,340 (int32 max ~ 2.1e9, n^2 overflows)
+        order = np.argsort(diag_idx.astype(np.int64) * np.int64(n) + positions.astype(np.int64))
 
         lines = np.zeros(n_entries, dtype=numba.int32)
         n_lines = 0
